Backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
python-dotenv==1.0.1
twilio==9.1.0
python-multipart==0.0.9
aiohttp==3.9.5
